Add support for Mac + fixes
Here's a bunch of fixes and support for Mac. Eleven commits as follows:

 - On a Mac, you need to do some prebuild installs somewhat differently
 - Based on the wiki, this is a sample configuration file complete with description
 - On an Apple Mac (Intel or Apple CPU) you need some more includes
 - min_elevation is an example where double or int is acceptable. The frequency value is now uint64
 - fd needed to be set in order for the code to correctly pass compiler error
 - verbose -v flag code was added - more to do. The return from sig_register() was made more verbose - not that it should ever be called
 - json_get_uint64_by_key() added to read frequency from json and handle 10Ghz etc
 - uint64_t for frequency variable is needed because QO-100 Es'Hail2 runs at 10Ghz and that's more than 32 bits
 - limits.h is only on Linux
 - I needed a fake rotator to test code on the road
 - The .gitignore file stops the build directory from uploading to GitHub

+++ src/cmd.c
@@ -22,7 +22,7 @@ int pre_doit(observation_t *obs)
 	char freq[32] = { 0 };
 
 	if (obs->active)
-		snprintf(freq, sizeof(freq), "%d", obs->active->frequency);
+		snprintf(freq, sizeof(freq), "%llu", (long long)obs->active->frequency);
 
 	char *args[] = { DEF_CONF_PYTHON, (char *) obs->cfg->cmd_script, "pre_doit", freq, NULL };
 	return request_cmd(DEF_CONF_PYTHON, args);
